WoT restructure B3: utils/{row,column,positionCache} -> axisSizing/{rowUtils,columnUtils,positionCache}
