+++ handsontable/src/3rdparty/walkontable/src/axisSizing/positionCache.ts
