+++ handsontable/src/3rdparty/walkontable/src/calculator/axisCalculation.ts
@@ -1,4 +1,4 @@
-import type { PositionCache } from '../utils/positionCache';
+import type { PositionCache } from '../axisSizing/positionCache';
 import type { ViewportBaseCalculator } from './viewportBase';
 
 export interface AxisCalculatorContext extends ViewportBaseCalculator {
